fancy

+++ BedWars/src/main/java/net/thisisnico/lolz/bedwars/Game.java
@@ -11,18 +11,19 @@ import net.thisisnico.lolz.common.adapters.DatabaseAdapter;
 import net.thisisnico.lolz.common.database.Clan;
 import net.thisisnico.lolz.common.database.Database;
 import net.thisisnico.lolz.common.database.User;
-import org.bukkit.Bukkit;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.Sound;
+import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Item;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.potion.PotionEffect;
 
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
 
 public class Game {
 
@@ -35,6 +36,8 @@ public class Game {
     @Getter
     private static final ArrayList<ResourceGenerator> generators = new ArrayList<>();
 
+    private static final HashMap<ArmorStand, Location> tpArmorStandsBack = new HashMap<>();
+
     private static boolean isRunning = false;
 
     @SuppressWarnings("BooleanMethodIsAlwaysInverted")
@@ -44,7 +47,16 @@ public class Game {
 
     public static boolean isPlayerInGame(Player player) {
         for (Team team : teams) {
-            if (team.getPlayers().contains(player)) {
+            if (team.getPlayers().contains(player.getName())) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public static boolean isOfflinePlayerInGame(OfflinePlayer player) {
+        for (Team team : teams) {
+            if (team.getPlayers().stream().anyMatch(name -> name.equalsIgnoreCase(player.getName()))) {
                 return true;
             }
         }
@@ -63,13 +75,25 @@ public class Game {
                     return false;
                 }
 
+                var destroyTeam = Game.getTeam(p);
+                if (destroyTeam == null) return false;
                 team.destroyBed();
-                broadcast("Кровать команды " + team.getName() + " была уничтожена игроком " + p.getName() + "!");
+                Bukkit.broadcast(Component.color("&f&lУНИЧТОЖЕНИЕ КРОВАТИ > &7Кровать команды ")
+                        .append(Component.color(team.getName()).color(team.getColor().getColor()))
+                        .append(Component.color(" &7была уничтожена игроком "))
+                        .append(Component.color(p.getName()).color(destroyTeam.getColor().getColor())));
                 team.setCoolDudeWhoBrokeDaBed(p);
+
+                // remove offline players from team
+                team.getPlayers().removeIf(player -> Bukkit.getPlayerExact(player) == null);
+                if (team.getPlayers().size() == 0) eliminateTeam(team);
 
                 var clan = DatabaseAdapter.getClan(p);
                 if (clan == null) p.sendMessage(Component.color("&cТы не в клане"));
-                else clan.givePoints(Const.POINTS_FOR_BED);
+                else {
+                    destroyTeam.addScore(Const.POINTS_FOR_BED);
+                    clan.givePoints(Const.POINTS_FOR_BED);
+                }
 
                 return true;
             }
@@ -123,6 +147,15 @@ public class Game {
 
     public static void start() {
         if (isRunning) return;
+
+        WorldEditStuff.load("arena");
+
+        for (Item entity : arena.getWorld().getEntitiesByClass(Item.class)) {
+            if (entity.getItemStack().getType().name().contains("BED")) {
+                entity.remove();
+            }
+        }
+
         var clans = new ArrayList<Clan>();
         Database.getClans().find().forEach(clans::add);
 
@@ -132,6 +165,13 @@ public class Game {
         for (Player player : Bukkit.getOnlinePlayers()) {
             var user = users.stream().filter(u -> u.getName().equalsIgnoreCase(player.getName())).findFirst().orElse(null);
             assert user != null;
+
+            if (user.isAdmin()) {
+                player.setGameMode(GameMode.SPECTATOR);
+                player.teleport(arena.getSpectatorSpawnLocation());
+                continue;
+            }
+
             Clan clan = null;
             if (user.hasClan()) clan = clans.stream().filter(c -> c.getTag().equalsIgnoreCase(user.getClan())).findFirst().orElse(null);
             if (clan == null) {
@@ -140,6 +180,9 @@ public class Game {
             }
 
             player.setFoodLevel(20);
+            player.setHealth(20);
+            player.setGameMode(GameMode.SURVIVAL);
+            player.getInventory().setContents(new ItemStack[] {});
 
             Clan finalClan = clan;
             var team = teams.stream().filter(t -> t.getName().equalsIgnoreCase(finalClan.getTag())).findFirst().orElse(null);
@@ -149,27 +192,35 @@ public class Game {
             }
 
             team.addPlayer(player);
+            player.teleport(team.getSpawnLocation());
         }
 
-        for (Team team : teams) {
-            for (Player player : team.getPlayers()) {
-                player.teleport(team.getSpawnLocation());
+        for (ArmorStand entity : arena.getWorld().getEntitiesByClass(ArmorStand.class)) {
+            if (!entity.getScoreboardTags().contains("generator")) {
+                tpArmorStandsBack.put(entity, entity.getLocation());
+                entity.teleport(new Location(entity.getWorld(), 0, 0, 0));
             }
         }
 
         isRunning = true;
+
+        for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
+            Coloring.updateColors(onlinePlayer);
+        }
     }
 
     public static void dispose() {
         for (Team team : teams) {
-            for (Player player : team.getPlayers()) {
-                player.teleport(arena.getWorld().getSpawnLocation());
+            for (String player : team.getPlayers()) {
+                if (Bukkit.getPlayerExact(player) != null)
+                    Bukkit.getPlayerExact(player).teleport(arena.getWorld().getSpawnLocation());
             }
         }
         teams.clear();
         for (ResourceGenerator generator : generators) {
-            generator.getHologram().remove();
+            generator.dispose();
         }
+        generators.clear();
         for (Block playerBlock : arena.getPlayerBlocks()) {
             playerBlock.setType(Material.AIR);
         }
@@ -181,18 +232,139 @@ public class Game {
                 entity.remove();
             }
         }
+
+        for (Map.Entry<ArmorStand, Location> entry : tpArmorStandsBack.entrySet()) {
+            entry.getKey().teleport(entry.getValue());
+        }
+        tpArmorStandsBack.clear();
+    }
+    
+    public static void kill(Player player, boolean respawn) {
+        Bukkit.getScheduler().runTaskLater(BedWars.getInstance(), () -> {
+            player.setGameMode(GameMode.SPECTATOR);
+            player.teleport(Game.getArena().getSpectatorSpawnLocation());
+        }, 1L);
+
+        Team team = Game.getTeam(player);
+
+        if (team == null) return;
+
+        OfflinePlayer killer = player.getKiller();
+        if (killer == null) killer = team.getCoolDudeWhoBrokeDaBed();
+
+        if (killer == null) Bukkit.broadcast(Component.color(player.getName()).color(team.getColor().getColor())
+                .append(Component.color(" &7умер"))
+                .append(Component.color(team.isBedDestroyed() ? " &b&lФИНАЛЬНОЕ УБИЙСТВО!" : "")));
+        else Bukkit.broadcast(Component.color(player.getName()).color(team.getColor().getColor())
+                .append(Component.color(" &7был убит игроком "))
+                .append(Component.color(killer.getName()).color(Game.getTeam(killer).getColor().getColor()))
+                .append(Component.color(team.isBedDestroyed() ? " &b&lФИНАЛЬНОЕ УБИЙСТВО!" : "")));
+
+        player.getInventory().clear();
+        player.setAllowFlight(true);
+        player.setFlying(true);
+
+        if (team.isBedDestroyed()) {
+            if (killer != null) {
+                var clan = DatabaseAdapter.getClan(killer);
+                if (clan == null) {
+                    if (killer.isOnline()) Objects.requireNonNull(killer.getPlayer()).sendMessage(Component.color("&cТы не в клане"));
+                    return;
+                }
+                clan.givePoints(Const.POINTS_FOR_FINAL_KILL);
+                Game.getTeam(killer).addScore(Const.POINTS_FOR_FINAL_KILL);
+            }
+
+            team.getPlayers().remove(player.getName());
+            eliminateTeam(team);
+        }
+        else if (respawn) respawn(player, 1);
+
+        for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
+            Coloring.updateColors(onlinePlayer);
+        }
+    }
+
+    public static void eliminateTeam(Team team) {
+        var size_only_online_players = team.getPlayers().stream()
+                .filter(player -> Bukkit.getPlayerExact(player) != null).count();
+        if (size_only_online_players == 0) {
+            Bukkit.broadcast(Component.color("&f&lУНИЧТОЖЕНИЕ КОМАНДЫ > ").append(Component.color(team.getName()).color(team.getColor().getColor())
+                    .append(Component.color(" &cвыбывает из игры!"))));
+            teams.remove(team);
+        }
+
+        if (teams.size() == 1) {
+            for (String offline : teams.get(0).getPlayers()) {
+                if (Bukkit.getPlayerExact(offline) != null) {
+                    Bukkit.getPlayerExact(offline).getInventory().setArmorContents(new ItemStack[] {});
+                }
+            }
+            Bukkit.broadcast(Component.color("&f"));
+            Bukkit.broadcast(Component.color("&9&m========================="));
+            Bukkit.broadcast(Component.color("&f"));
+            Bukkit.broadcast(Component.color(teams.get(0).getName()).color(teams.get(0).getColor().getColor())
+                    .append(Component.color(" &7 - &b&lПОБЕДИТЕЛЬ")));
+            Bukkit.broadcast(Component.center("     &d(+"+(Const.POINTS_FOR_WIN + teams.get(0).getScore())+" очков)"));
+            Bukkit.broadcast(Component.color("&f"));
+            Bukkit.broadcast(Component.color("&9&m========================="));
+            Bukkit.broadcast(Component.color("&f"));
+            Clan.get(teams.get(0).getName()).givePoints(Const.POINTS_FOR_WIN);
+            stop();
+        }
+    }
+
+    public static void respawn(Player player, int k) {
+        Team team = Game.getTeam(player);
+        if (team == null) return;
+        if (team.isBedDestroyed()) {
+            player.setGameMode(GameMode.SPECTATOR);
+            player.teleport(Game.getArena().getSpectatorSpawnLocation());
+            player.sendMessage(Component.color("&cКровать твоей команды была сломана. Ты не возродишься."));
+            return;
+        }
+        final int[] i = {Const.RESPAWN_DELAY * k};
+        Bukkit.getScheduler().runTaskTimer(BukkitUtils.getPlugin(), task -> {
+            if (!player.isOnline()) {
+                task.cancel();
+                return;
+            }
+            if (i[0] == 0) {
+                player.teleport(team.getSpawnLocation());
+                player.setGameMode(GameMode.SURVIVAL);
+                player.setAllowFlight(false);
+                player.setFlying(false);
+                player.getInventory().setContents(new ItemStack[]{});
+                player.setHealth(20);
+                player.setFoodLevel(20);
+                player.setSaturation(10);
+                for (PotionEffect activePotionEffect : player.getActivePotionEffects()) {
+                    player.removePotionEffect(activePotionEffect.getType());
+                }
+                task.cancel();
+            } else {
+                player.teleport(Game.getArena().getSpectatorSpawnLocation());
+                player.setGameMode(GameMode.SPECTATOR);
+                player.sendActionBar(Component.color("&cРеспавн через " + i[0] + " сек"));
+                i[0]--;
+            }
+        }, 0L, 20L);
     }
 
     public static void stop() {
         isRunning = false;
         TeamColor.clearTakenColors();
 
-        // TODO. end da game
+        for (Player onlinePlayer : Bukkit.getOnlinePlayers()) {
+            Coloring.updateColors(onlinePlayer);
+        }
+
+        dispose();
     }
 
-    public static Team getTeam(Player player) {
+    public static Team getTeam(OfflinePlayer player) {
         for (Team team : teams) {
-            if (team.getPlayers().contains(player)) {
+            if (team.getPlayers().contains(player.getName())) {
                 return team;
             }
         }
